More frontend work, mainly on create novel modal

+++ src/frontend/my-novel-builder/src/app/services/novel.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Observable, map } from 'rxjs';
 import { NovelDto } from '../types/dtos/novel/novel.dto';
 import { environment } from '../../environment';
-import { mockObservable, mockedNovels } from './mock';
+import { mockObservable, mockedNovel, mockedNovels } from './mock';
 import { Injectable } from '@angular/core';
 import { NovelCoverImageDto } from '../types/dtos/novel/novel-cover-image.dto';
 import { CreateNovelDto } from '../types/dtos/novel/create-novel.dto';
@@ -23,16 +23,6 @@ export class NovelService {
       : this.http.get<NovelDto[]>(`${this.baseUrl}/novels`);
   }
 
-  getNovelCoverImageUrl(novelId: string): Observable<string | null> {
-    return this.mocked
-      ? mockObservable(`https://picsum.photos/seed/${novelId}/200/300`)
-      : this.http
-          .get<NovelCoverImageDto>(
-            `${this.baseUrl}/novel/${novelId}/cover-image`
-          )
-          .pipe(map((coverImage) => coverImage?.location ?? null));
-  }
-
   uploadNovelCoverImage(
     novelId: string,
     file: File
@@ -40,21 +30,31 @@ export class NovelService {
     const formData = new FormData();
     formData.append('file', file);
 
-    return this.http.post<NovelCoverImageDto>(
-      `${this.baseUrl}/novel/${novelId}/cover-image`,
-      formData
-    );
+    return this.mocked
+      ? mockObservable({ id: 'mocked', location: '' })
+      : this.http.post<NovelCoverImageDto>(
+          `${this.baseUrl}/novel/${novelId}/cover-image`,
+          formData
+        );
   }
 
   createNovel(novel: CreateNovelDto): Observable<NovelDto> {
-    return this.http.post<NovelDto>(`${this.baseUrl}/novel`, novel);
+    return this.mocked
+      ? mockObservable(mockedNovel)
+      : this.http.post<NovelDto>(`${this.baseUrl}/novel`, novel);
   }
 
   updateNovel(novel: UpdateNovelDto): Observable<NovelDto> {
-    return this.http.put<NovelDto>(`${this.baseUrl}/novel`, novel);
+    return this.mocked
+      ? mockObservable(mockedNovel)
+      : this.http.put<NovelDto>(`${this.baseUrl}/novel`, novel);
   }
 
   deleteNovel(novelId: string): Observable<void> {
+    if (this.mocked) {
+      return mockObservable<void>(undefined);
+    }
+
     return this.http.delete<void>(`${this.baseUrl}/novel/${novelId}`);
   }
 }
